all register controller bugs solved

+++ src/utils/cloudinary.js
@@ -2,9 +2,9 @@ import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
 
 cloudinary.config({ 
-    cloud_name: process.env.CLOUDINAR_CLOUD_NAME, 
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
     api_key: process.env.CLOUDINARY_API_KEY, 
-    api_secret: process.env.CLOUDINARY_API_SECRET
+    api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
@@ -15,11 +15,13 @@ const uploadOnCloudinary = async (localFilePath) => {
             resource_type: "auto"
         })
         
-        console.log("File uploaded on cloudinary sccessfully", response.url)
+        fs.unlinkSync(localFilePath);
         return response;
 
-    } catch (error) {
-        fs.unlinkSync(localFilePath)
+    } catch (error) {       
+        if(localFilePath && fs.existsSync(localFilePath)){
+            fs.unlinkSync(localFilePath)
+        }  
         return null
     }
 }
